fix some comments;
