+++ app/src/pages/Desktop/Protocols/ProtocolVisualization/consants.ts
@@ -1,0 +1,10 @@
+export const POTENTIAL_TRASH_COMMAND_TYPES = [
+  'moveToAddressableArea',
+  'moveToAddressableAreaForDropTip',
+  'dropTip',
+  'dropTipInPlace',
+  'airGapInPlace',
+  'blowOutInPlace',
+  'blowOut',
+  'airGap',
+]
